IReadFile::read now returning size_t (like fread in c-lib) instead of s32. Also sizeToRead parameter changed from u32 to size_t. (corresponding change to IWriteFile::write is planned for next days)
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@5333 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/IReadFile.h
@@ -21,7 +21,7 @@ namespace io
 		/** \param buffer Pointer to buffer where read bytes are written to.
 		\param sizeToRead Amount of bytes to read from the file.
 		\return How many bytes were read. */
-		virtual s32 read(void* buffer, u32 sizeToRead) = 0;
+		virtual size_t read(void* buffer, size_t sizeToRead) = 0;
 
 		//! Changes position in file
 		/** \param finalPos Destination position in the file.
