+++ source/Irrlicht/CMemoryFile.cpp
@@ -28,9 +28,9 @@ CMemoryReadFile::~CMemoryReadFile()
 
 
 //! returns how much was read
-s32 CMemoryReadFile::read(void* buffer, u32 sizeToRead)
+size_t CMemoryReadFile::read(void* buffer, size_t sizeToRead)
 {
-	s32 amount = static_cast<s32>(sizeToRead);
+	long amount = static_cast<long>(sizeToRead);
 	if (Pos + amount > Len)
 		amount -= Pos + amount - Len;
 
@@ -42,7 +42,7 @@ s32 CMemoryReadFile::read(void* buffer, u32 sizeToRead)
 
 	Pos += amount;
 
-	return amount;
+	return static_cast<size_t>(amount);
 }
 
 //! changes position in file, returns true if successful
