+++ source/Irrlicht/CReadFile.cpp
@@ -29,12 +29,12 @@ CReadFile::~CReadFile()
 
 
 //! returns how much was read
-s32 CReadFile::read(void* buffer, u32 sizeToRead)
+size_t CReadFile::read(void* buffer, size_t sizeToRead)
 {
 	if (!isOpen())
 		return 0;
 
-	return (s32)fread(buffer, 1, sizeToRead, File);
+	return fread(buffer, 1, sizeToRead, File);
 }
